Excel and Word update
When converting Excel, we allow a maximum number of rows to
be set, after which conversion will not be attempted because
it makes Excel go *spoing*. Using UsedRange seems to include
rows that are blank, so change the code to work out the maximum
row based on the last non-blank row.

When converting Word, make sure that we avoid auto-saving
the normal template.