Try and handle cases where embedded fonts block printing [Issue #15]
When a presentation contains embedded, read-only fonts, the file
will open as read-only, and the save and export interop calls will
throw exceptions. This change attempts to detect when the object
model is not available and work-around the save as problem by
inserting slides from the original file into a new presentation.

This may result in an odd looking conversion, as the slide masters
from the original file won't come across. Then again, if the embedded
fonts are missing, the layout is going to be somewhat odd anyway.