Be careful when looking through headers/footers
When running through the headers and footers to update fields,
Word interop will cheerfully initialise a header and footer
in a section even if it was empty. This can lead to un-expected
whitespace appearing in documents because of the newly created
header/footer.

What this change does is set the header/footer distance to zero
where the header/footer doesn't contain any shapes.

See https://officetopdf.codeplex.com/discussions/638088