Rejig how field updating is performed
Previously, when we updated fields in stories and headers/footers,
not everything was cleaned up, which meant Word could stay
in the header/footer view. This could introduce additional whitespace
onto the resulting PDF.

Now the update code has been refactored to hopefully remove this
problem.

Also, try not to alert if Word is not the default application.