ResolveOxid.
These code changes add a new C++ assembly that allows the calling
of the ResolveOxid function based on the idl.

The same code could not be part of the managed C++ assembly.

The code is currently delivering the strings that need to be parsed
to get the port. From the port then hopefully the process id can
be found.