Watchdog.
These code changes add the Watchdog class and an associated factory.

The watchdog kills the COM server process after a specific time period in seconds.
If an invalid timeout period is used then a NullWatchdog is returned which implements
the IWatchdog service and can be used without the need for null checks in the client
code.